# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "block-funders",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api"
        }
    ],
    "paths": {
...
        "/permissions": {
            "get": {
                "tags": [
                    "Permissions"
                ],
                "description": "get all permissions",
                "operationId": "d1e098c4978abaa9609320b949d0cd0e",
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden. This action is unauthorized."
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/permissions/me": {
            "get": {
                "tags": [
                    "Permissions"
                ],
                "description": "Get my permissions",
                "operationId": "615e8497a365bb57653dbea6eccd0500",
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden. This action is unauthorized."
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/roles/{role}/permissions": {
            "get": {
                "tags": [
                    "Permissions"
                ],
                "summary": "Get permissions for a role",
                "operationId": "dfd082a4d90cb1c51bca7fcda355a1e8",
                "parameters": [
                    {
                        "name": "role",
                        "in": "path",
                        "description": "The ID of the role",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden. This action is unauthorized."
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Permissions"
                ],
                "summary": "Set permissions for a role",
                "operationId": "a0547786d6ee456ec30406eed2364a61",
                "parameters": [
                    {
                        "name": "role",
                        "in": "path",
                        "description": "The ID of the role",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "permissions[0]": {
                                        "description": "The names of the permissions",
                                        "type": "string"
                                    },
                                    "permissions[1]": {
                                        "description": "The names of the permissions",
                                        "type": "string"
                                    },
                                    "permissions[2]": {
                                        "description": "The names of the permissions",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden. This action is unauthorized."
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearer_token": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}